add python file to replay raw recorded bytes
permits to test the ground station software using old flight data.
The python script reads the lines from a bytes_received file, and it writes them to a serial port.
To complete the setup, the use of a serial bridge is recommended: for example COM1 <-> COM14
The groundstation software reads COM1, while this script writes to COM14